Removing `NVMBaseApi` abstract class
